database implementation: improved the layer inbetween databse and service
* added the creation of objects from database cursor and
viceversa
* added a method to delete all the routes and all the route
points that belong to the route

+++ tracklib/src/main/java/com/ant/track/lib/db/content/TrackMeDatabaseUtils.java
@@ -24,7 +24,7 @@ public interface TrackMeDatabaseUtils {
      *
      * @param id of the route to be deleted.
      */
-    void deleteRouteTrack(int id);
+    void deleteRouteTrack(long id);
 
     /**
      * updates a route from the database.
